Ensure udhcpc is not running for statically defined eth1 interface.
This causes issues where the udhcpc process may eventually attempt to
acquire a new IP address from VitualBox or VMware's DHCP server, and
thus severing the connection to the Docker daemon (from the client's
point of view).

Sadly this is more of a Tiny Core Linux boot issue with the
busybox/udhcpc startup and not in boot2docker so it'll need to be
handled in Vagrant middleware.

Closes #28
Closes #24
Closes #8